# server/openapi-spec.yaml
openapi: '3.0.3'
info:
  description: Server to support Ver-ID in-browser face recognition
  version: '1.0.0'
  title: Ver-ID API Server
  contact:
    email: support@appliedrecognition.com
servers:
  - url: https://id-check.ver-id.com/
    description: Demo
paths:
...
  '/demo':
    get:
      tags:
        - Demo
      summary: Demo page featuring ID capture, live face capture and face comparison
      description: >-
        The page guides the user to capture their ID card before it proceeds to capture the user's face. 
        Finally, the face on the ID card is compared to the live face and the page outputs a "Pass" or a "Warning".
      responses:
        '200':
          description: OK
          content:
            'text/html':
              schema:
                type: string
        '405':
          $ref: '#/components/responses/FeatureDisabledResponse'
  '/detectFace':
    post:
      tags:
        - Face Detection
      summary: Detect a face in an image
      requestBody:
        $ref: '#/components/requestBodies/FaceDetectionRequest'
      responses:
        '200':
          $ref: '#/components/responses/FaceDetectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  '/detectIdCard':
    post:
      tags:
        - ID Card Detection
      summary: Detect an ID card in an image
      requestBody:
        $ref: '#/components/requestBodies/IDCardDetectionRequest'
      responses:
        '200':
          $ref: '#/components/responses/IDCardDetectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/FeatureDisabledResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  '/compareFaces':
    post:
      tags:
        - Face Recognition
      summary: Compare two face recognition templates
      description: Returns a similarity score
      requestBody:
        $ref: '#/components/requestBodies/CompareFacesRequest'
      responses:
        '200':
          $ref: '#/components/responses/CompareFacesResponse'
        '400':
          description: Bad request, e.g., missing strings array or array wrong length
          content:
            'text/plain':
              schema:
                type: string
                example: >-
                  Specify exactly two face templates to compare
        '406':
          description: Request cannot accept application/json response
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  requestBodies:
    FaceDetectionRequest:
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/FaceDetectionRequest'
          examples:
            idCard:
              summary: Detect a face on the front of an ID card
              value:
                image: data:image/jpeg;base64,AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcY
                calculate_authenticity_score: true
            image:
              summary: Detect a face in an image
              value:
                image: data:image/jpeg;base64,MDQ4PEBESExQVFhcYAQIDBAUGBwgJCgs                
        'multipart/form-data':
          schema:
            $ref: '#/components/schemas/FaceImageUpload'
          encoding:
            image:
              contentType: image/jpeg, image/png
    IDCardDetectionRequest:
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/IDCardDetectionRequest'
        'multipart/form-data':
          schema:
            $ref: '#/components/schemas/IDCardImageUpload'
          encoding:
            front:
              contentType: image/jpeg, image/png
            back:
              contentType: image/jpeg, image/png
    CompareFacesRequest:
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/CompareFacesRequest'
  responses:
    FaceDetectionResponse:
      description: OK
      content:
        'application/json':
          schema:
            allOf:
              - $ref: '#/components/schemas/Face'
              - type: object
                properties:
                  authenticityScores:
                    description: Map of authenticity scores returned by the models prefixed with the object keys
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/AuthenticityScore'
                    example:
                      license01: 0.983
                      license: 0.877
    IDCardDetectionResponse:
      description: OK
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/ScanResult'
    CompareFacesResponse:
      description: OK
      content:
        'text/plain':
          schema:
            type: string
            example: 3.45
        'application/json':
          schema:
            type: object
            properties:
              score:
                type: number
                format: double
                example: 3.45
    FeatureDisabledResponse:
      description: Method not allowed
      content:
        'text/plain':
          schema:
            type: string
            description: The method is not allowed because the feature is disabled on this server.
    BadRequest:
      description: Bad request
      content:
        'text/plain':
          schema:
            type: string
            description: Description of the failure, e.g., which properties are missing
            example: Missing image
    ServerError:
      description: Server error
      content:
        'text/plain':
          schema:
            type: string
            description: Error description
            example: Unable to load server configuration
  schemas:
    FaceDetectionRequest:
      type: object
      properties:
        image:
          $ref: '#/components/schemas/ImageDataURL'
        calculate_authenticity_score:
          type: boolean
          description: Set to `true` to calculate authenticity score (suitable for images of ID cards)
      required:
        - image
    IDCardDetectionRequest:
      type: object
      properties:
        front:
          $ref: '#/components/schemas/ImageDataURL'
        back:
          $ref: '#/components/schemas/ImageDataURL'
      minProperties: 1
    FaceImageUpload:
      type: object
      properties:
        image:
          $ref: '#/components/schemas/Image'
        calculate_authenticity_score:
          type: boolean
      required:
        - image
    IDCardImageUpload:
      type: object
      properties:
        front:
          $ref: '#/components/schemas/Image'
        back:
          $ref: '#/components/schemas/Image'
        iin_safelist:
          type: array
          description: Array of acceptable issuer identification numbers IINs
          items:
            $ref: '#/components/schemas/IIN'
        iin_blocklist:
          type: array
          description:  Array of unacceptable issuer identification numbers IINs
          items:
            $ref: '#/components/schemas/IIN'
      minProperties: 1
    CompareFacesRequest:
      type: array
      items:
        description: Face recognition template
        type: string
        format: base64
        example: ZmFjZVRlbXBsYXRlCg==
      minItems: 2
      example:
        - MWZhY2UxVGVtcGxhdGVPbmUK
        - MmZhY2UyVGVtcGxhdGVUd28K
...
    BlinkIdResult:
      type: object
      properties:
        dateOfBirth:
          $ref: '#/components/schemas/Date'
        classInfo:
          type: object
          properties:
            country:
              type: string
            region:
              type: string
            type:
              type: string
            countryName:
              type: string
            isoAlpha3CountryCode:
              type: string
            isoAlpha2CountryCode:
              type: string
            isoNumericCountryCode:
              type: string
        type:
          type: string
        isBelowAgeLimit:
          type: boolean
        age:
          type: number
        recognitionStatus:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        address:
          type: string
        dateOfIssue:
          $ref: '#/components/schemas/Date'
        dateOfExpiry:
          $ref: '#/components/schemas/Date'
        documentNumber:
          type: string
        sex:
          type: string
        driverLicenseDetailedInfo:
          $ref: '#/components/schemas/DriverLicenseDetailedInfo'
        fullDocumentImageBase64:
          type: string
        faceImageBase64:
          type: string
        additionalNameInformation:
          type: string
        additionalAddressInformation:
          type: string
        placeOfBirth:
          type: string
        nationality:
          type: string
        race:
          type: string
        religion:
          type: string
        profession:
          type: string
        maritalStatus:
          type: string
        residentialStatus:
          type: string
        employer:
          type: string
        personalIdNumber:
          type: string
        documentAdditionalNumber:
          type: string
        documentOptionalAdditionalNumber:
          type: string
        issuingAuthority:
          type: string
        mrzData:
          type: object
          properties:
            rawMrzString:
              type: string
            documentCode:
              type: string
            issuer:
              type: string
            documentNumber:
              type: string
            opt1:
              type: string
            opt2:
              type: string
            gender:
              type: string
            nationality:
              type: string
            primaryId:
              type: string
            secondaryId:
              type: string
            alienNumber:
              type: string
            applicationReceiptNumber:
              type: string
            immigrantCaseNumber:
              type: string
            mrzVerified:
              type: boolean
            mrzParsed:
              type: boolean
            dateOfBirth:
              $ref: '#/components/schemas/Date'
            dateOfExpiry:
              $ref: '#/components/schemas/Date'
            documentType:
              type: string
            issuerName:
              type: string
            nationalityName:
              type: string
        conditions:
          type: string
        localizedName:
          type: string
        dateOfExpiryPermanent:
          type: boolean
        additionalPersonalIdNumber:
          type: string
        viz:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            fullName:
              type: string
            additionalNameInformation:
              type: string
            localizedName:
              type: string
            address:
              type: string
            additionalAddressInformation:
              type: string
            placeOfBirth:
              type: string
            nationality:
              type: string
            race:
              type: string
            religion:
              type: string
            profession:
              type: string
            maritalStatus:
              type: string
            residentialStatus:
              type: string
            employer:
              type: string
            sex:
              type: string
            dateOfBirth:
              $ref: '#/components/schemas/Date'
            dateOfIssue:
              $ref: '#/components/schemas/Date'
            dateOfExpiry:
              $ref: '#/components/schemas/Date'
            dateOfExpiryPermanent:
              type: boolean
            documentNumber:
              type: string
            personalIdNumber:
              type: string
            documentAdditionalNumber:
              type: string
            additionalPersonalIdNumber:
              type: string
            documentOptionalAdditionalNumber:
              type: string
            issuingAuthority:
              type: string
            driverLicenseDetailedInfo:
              $ref: '#/components/schemas/DriverLicenseDetailedInfo'
            conditions:
              type: string
        barcode:
          type: object
          properties:
            rawDataBase64:
              type: string
            stringData:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            middleName:
              type: string
            fullName:
              type: string
            additionalNameInformation:
              type: string
            address:
              type: string
            placeOfBirth:
              type: string
            nationality:
              type: string
            race:
              type: string
            religion:
              type: string
            profession:
              type: string
            maritalStatus:
              type: string
            residentialStatus:
              type: string
            employer:
              type: string
            sex:
              type: string
            dateOfBirth:
              $ref: '#/components/schemas/Date'
            dateOfIssue:
              $ref: '#/components/schemas/Date'
            dateOfExpiry:
              $ref: '#/components/schemas/Date'
            documentNumber:
              type: string
            personalIdNumber:
              type: string
            documentAdditionalNumber:
              type: string
            issuingAuthority:
              type: string
            addressDetailedInfo:
              type: object
              properties:
                street:
                  type: string
                postalCode:
                  type: string
                city:
                  type: string
                jurisdiction:
                  type: string
            driverLicenseDetailedInfo:
              $ref: '#/components/schemas/DriverLicenseDetailedInfo'
            extendedElements:
              type: array
              items:
                type: string
        imageAnalysisResult:
          type: object
          properties:
            blurred:
              type: boolean
            documentImageColorStatus:
              type: string
            documentImageMoireStatus:
              type: string
            faceDetectionStatus:
              type: string
            mrzDetectionStatus:
              type: string
            barcodeDetectionStatus:
              type: string
        processingStatus:
          type: string
        recognitionMode:
          type: string
        signatureImageBase64:
          type: string
      example:
        dateOfBirth:
          day: 18
          month: 3
          year: 1990
          successfullyParsed: true
          originalString: 03/18/1990
        classInfo:
          country: COUNTRY_USA
          region: REGION_FLORIDA
          type: TYPE_DL
          countryName: UNITED STATES OF AMERICA
          isoAlpha3CountryCode: USA
          isoAlpha2CountryCode: US
          isoNumericCountryCode: '840'
        type: BLINK_ID
        isBelowAgeLimit: null
        age: -1
        recognitionStatus: VALID
        firstName: SAMPLE
        lastName: SPECIMEN
        fullName: ''
        address: |-
          123 MAIN STREET APT. 1
          ORLANDO, FL 17101-0000
        dateOfIssue:
          day: 18
          month: 3
          year: 2018
          successfullyParsed: true
          originalString: 03/18/2018
        dateOfExpiry:
          day: 18
          month: 3
          year: 2025
          successfullyParsed: true
          originalString: 03/18/2025
        documentNumber: S123-456-78-123-0
        sex: M
        driverLicenseDetailedInfo:
          restrictions: NONE
          endorsements: NONE
          vehicleClass: C
          conditions: ''
        fullDocumentImageBase64: ''
        faceImageBase64: ''
        additionalNameInformation: ''
        additionalAddressInformation: ''
        placeOfBirth: ''
        nationality: ''
        race: ''
        religion: ''
        profession: ''
        maritalStatus: ''
        residentialStatus: ''
        employer: ''
        personalIdNumber: ''
        documentAdditionalNumber: ''
        documentOptionalAdditionalNumber: ''
        issuingAuthority: ''
        mrzData:
          rawMrzString: ''
          documentCode: ''
          issuer: ''
          documentNumber: ''
          opt1: ''
          opt2: ''
          gender: ''
          nationality: ''
          primaryId: ''
          secondaryId: ''
          alienNumber: ''
          applicationReceiptNumber: ''
          immigrantCaseNumber: ''
          mrzVerified: false
          mrzParsed: false
          dateOfBirth:
            day: 0
            month: 0
            year: 0
            successfullyParsed: false
            originalString: ''
          dateOfExpiry:
            day: 0
            month: 0
            year: 0
            successfullyParsed: false
            originalString: ''
          documentType: UNKNOWN
          issuerName: ''
          nationalityName: ''
        conditions: ''
        localizedName: ''
        dateOfExpiryPermanent: false
        additionalPersonalIdNumber: ''
        viz:
          firstName: SAMPLE
          lastName: SPECIMEN
          fullName: ''
          additionalNameInformation: ''
          localizedName: ''
          address: |-
            123 MAIN STREET APT. 1
            ORLANDO, FL 17101-0000
          additionalAddressInformation: ''
          placeOfBirth: ''
          nationality: ''
          race: ''
          religion: ''
          profession: ''
          maritalStatus: ''
          residentialStatus: ''
          employer: ''
          sex: M
          dateOfBirth:
            day: 18
            month: 3
            year: 1990
            successfullyParsed: true
            originalString: 03/18/1990
          dateOfIssue:
            day: 18
            month: 3
            year: 2018
            successfullyParsed: true
            originalString: 03/18/2018
          dateOfExpiry:
            day: 18
            month: 3
            year: 2025
            successfullyParsed: true
            originalString: 03/18/2025
          dateOfExpiryPermanent: false
          documentNumber: S123-456-78-123-0
          personalIdNumber: ''
          documentAdditionalNumber: ''
          additionalPersonalIdNumber: ''
          documentOptionalAdditionalNumber: ''
          issuingAuthority: ''
          driverLicenseDetailedInfo:
            restrictions: NONE
            endorsements: NONE
            vehicleClass: C
            conditions: ''
          conditions: ''
        barcode:
          rawDataBase64: ''
          stringData: ''
          firstName: ''
          lastName: ''
          middleName: ''
          fullName: ''
          additionalNameInformation: ''
          address: ''
          placeOfBirth: ''
          nationality: ''
          race: ''
          religion: ''
          profession: ''
          maritalStatus: ''
          residentialStatus: ''
          employer: ''
          sex: ''
          dateOfBirth:
            day: 0
            month: 0
            year: 0
            successfullyParsed: false
            originalString: ''
          dateOfIssue:
            day: 0
            month: 0
            year: 0
            successfullyParsed: false
            originalString: ''
          dateOfExpiry:
            day: 0
            month: 0
            year: 0
            successfullyParsed: false
            originalString: ''
          documentNumber: ''
          personalIdNumber: ''
          documentAdditionalNumber: ''
          issuingAuthority: ''
          addressDetailedInfo:
            street: ''
            postalCode: ''
            city: ''
            jurisdiction: ''
          driverLicenseDetailedInfo:
            restrictions: ''
            endorsements: ''
            vehicleClass: ''
            conditions: ''
          extendedElements: []
        imageAnalysisResult:
          blurred: false
          documentImageColorStatus: COLOR
          documentImageMoireStatus: NOT_AVAILABLE
          faceDetectionStatus: DETECTED
          mrzDetectionStatus: NOT_DETECTED
          barcodeDetectionStatus: NOT_DETECTED
        processingStatus: SUCCESS
        recognitionMode: FULL_RECOGNITION
        signatureImageBase64: ''
...
    ImageDataURL:
      type: string
      example: data:image/jpeg;base64,AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcY
    Image:
      type: string
      format: binary
    Date:
      type: object
      properties:
        day:
          type: integer
        month:
          type: integer
        year:
          type: integer
        originalString:
          type: string
      required: [day, month, year]
      example:
        day: 21
        month: 9
        year: 2020
        originalString: '2020/09/21'
    IIN:
      type: object
      properties:
        issuerIdentificationNumber:
          type: string
          example: 636012
    AuthenticityScore:
      type: number
      format: double
      minimum: 0
      maximum: 1
      example: 0.998
      description: Authenticity score (0.0 = ID is fake, 1.0 = ID is authentic)
    Face:
      type: object
      properties:
        jpeg:
          type: string
          format: base64
          description: Base64-encoded JPEG image of the detected face
        faceTemplate:
          type: string
          format: base64
          description: Base64-encoded face recognition template
    ScanResult:
      type: object
      properties:
        front:
          $ref: '#/components/schemas/BlinkIdResult'
        back:
          $ref: '#/components/schemas/UsdlResult'
      minProperties: 1